+++ lib/src/metrics/instruments/counter.dart
@@ -164,6 +164,7 @@ class Counter<T extends num> implements APICounter<T>, SDKInstrument {
       description: description,
       unit: unit,
       type: MetricType.sum,
+      isMonotonic: _storage.isMonotonic,
       points: points,
     );